* Implement threading of simple ai and lua ai * The above meant buffering a lot of stuff and applying it inbetween frames -   creations of units, deletions of them, events, new actions. * Don't allocate all movementType and traversalTime arrays for all combos   of movement types and size by default; create them dynamically   at runtime instead. * Rename a few movement types * Optimize movement a bit * Reduce memory usage slightly in many places. * Fixes in the Lua scripts * Add the aivsai level and its map * A small optimization in the binary heap code
git-svn-id: https://bloopz.zapto.org/svn/nightfall/projekt/trunk@954 d57568f3-fe0c-0410-b608-e86116ac1d00

+++ src/dimension.h
@@ -110,7 +110,7 @@ namespace Game
 			PlayerType        type;
 			vector<Unit*>     vUnits;
 			vector<Projectile*> vProjectiles;
-			int**             NumUnitsSeeingSquare;
+			Uint16**          NumUnitsSeeingSquare;
 			PlayerState*      states;
 			Resources         resources;
 			Resources         oldResources;
@@ -124,9 +124,9 @@ namespace Game
 		struct World
 		{
 			float**         ppHeight;
-			int**           ppSteepness;
+			Uint16**        ppSteepness;
 			float**         ppWater[3];
-			int**           NumLightsOnSquare;
+			Uint16**        NumLightsOnSquare;
 			vector<Unit*>   vUnits;
 			vector<UnitType*>   vUnitTypes;
 			vector<Player*> vPlayers;
